+++ eslint.config.mjs
@@ -18,7 +18,6 @@ export default [
     languageOptions: {
       globals: {
         ...globals.node,
-        ...globals.mocha,
       },
     },
   },